requirements.txt
# --- Jupyter + Utilities ---
jupyterlab
notebook
ipykernel
python-dotenv

# --- NLP ---
spacy
textstat
# download model separately: python -m spacy download en_core_web_sm

# --- Transformers & Embeddings ---
transformers>=4.21.0
torch>=1.12.0
sentence-transformers
openai
# DialoGPT-Medium specific dependencies
tokenizers>=0.13.0

# --- LangChain Ecosystem ---
langchain
langchain-openai
langchain-community
langgraph
chromadb

# --- Speech Recognition ---
faster-whisper
# openai-whisper   # optional slower baseline

# --- Text-to-Speech ---
pyttsx3
# TTS   # (optional: coqui.ai TTS, better quality)

# --- Backend ---
fastapi
uvicorn[standard]

# --- Frontend ---
streamlit
streamlit-webrtc

# --- Database & Storage ---
psycopg2-binary
sqlalchemy
boto3   # for S3
supabase

# --- Analytics & Visualization ---
plotly
seaborn
pandas
matplotlib
numpy

# --- Dev & Testing Tools ---
pytest
black
isort

# --- Additional Dependencies ---
pydantic
fastapi
